+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 
 ### FEATURES/ENHANCEMENTS:
+* ClientLists
+  * Added support for `REQUEST_HEADER_NAME_VALUE` type client lists.
 
 
 * Cloud Access